Split out EventReference, StateKeyTuple and EventBuilder
