+++ pdu.go
@@ -1,6 +1,9 @@
 package gomatrixserverlib
 
 import (
+	"encoding/json"
+	"fmt"
+
 	"github.com/matrix-org/gomatrixserverlib/spec"
 	"golang.org/x/crypto/ed25519"
 )
@@ -61,4 +64,57 @@ func ToPDUs[T PDU](events []T) []PDU {
 		result[i] = events[i]
 	}
 	return result
+}
+
+// A StateKeyTuple is the combination of an event type and an event state key.
+// It is often used as a key in maps.
+type StateKeyTuple struct {
+	// The "type" key of a matrix event.
+	EventType string
+	// The "state_key" of a matrix event.
+	// The empty string is a legitimate value for the "state_key" in matrix
+	// so take care to initialise this field lest you accidentally request a
+	// "state_key" with the go default of the empty string.
+	StateKey string
+}
+
+// An EventReference is a reference to a matrix event.
+type EventReference struct {
+	// The event ID of the event.
+	EventID string
+	// The sha256 of the redacted event.
+	EventSHA256 spec.Base64Bytes
+}
+
+// UnmarshalJSON implements json.Unmarshaller
+func (er *EventReference) UnmarshalJSON(data []byte) error {
+	var tuple []spec.RawJSON
+	if err := json.Unmarshal(data, &tuple); err != nil {
+		return err
+	}
+	if len(tuple) != 2 {
+		return fmt.Errorf("gomatrixserverlib: invalid event reference, invalid length: %d != 2", len(tuple))
+	}
+	if err := json.Unmarshal(tuple[0], &er.EventID); err != nil {
+		return fmt.Errorf("gomatrixserverlib: invalid event reference, first element is invalid: %q %v", string(tuple[0]), err)
+	}
+	var hashes struct {
+		SHA256 spec.Base64Bytes `json:"sha256"`
+	}
+	if err := json.Unmarshal(tuple[1], &hashes); err != nil {
+		return fmt.Errorf("gomatrixserverlib: invalid event reference, second element is invalid: %q %v", string(tuple[1]), err)
+	}
+	er.EventSHA256 = hashes.SHA256
+	return nil
+}
+
+// MarshalJSON implements json.Marshaller
+func (er EventReference) MarshalJSON() ([]byte, error) {
+	hashes := struct {
+		SHA256 spec.Base64Bytes `json:"sha256"`
+	}{er.EventSHA256}
+
+	tuple := []interface{}{er.EventID, hashes}
+
+	return json.Marshal(&tuple)
 }
